Move data loader and data sender to services
