+++ kiosk/src/js/application_updater.js
@@ -9,7 +9,7 @@ const NetworkErrorsHandlingService = require('./services/error-handlers/network_
 const AutoupdateScheduleWatcher = require('./services/data/autoupdate_schedule_watcher');
 const ConnectionStatusService = require('./services/network/connection_status_service');
 const Logger = require('./services/logger/logger');
-const DataSender = require('./data_sender');
+const DataSender = require('./services/data/data_sender');
 const DataStorage = require('./storage/data_storage');
 const LocalStorageManager = require('./helpers/local_storage_helper').LocalStorageManager;
 const StorageManager = require('./helpers/storage_manager');
